remove leftover partial file when wget fails
When wget fails, it still opens the output file, which will be left as at least a 0-byte file. On OBS, this can lead to "successfully" built RPMs with 0-byte files if the SOURCE files are not unpacked but just copied to their final location. Instead, remove whatever file is left over.